Handle generic power provider types & recharging batteries
By moving the generic power provider stuff for current and max power outside the 'if' block, it will function for all power providers, except for batteries in 'recharge' mode, which have a maxOutput value of 0 MW and their potential output needs to be read from the 'DetailedInfo' property as a special case.  Also differentiating between power producers that are enabled and able to function as current power, producers that are disabled but available to function as auxiliary, and producers that are unable to function as not relevant to the output.